# docs/openapi.yaml
# region OAS3
openapi: 3.0.3
info:
  title: memory-mall
  description: Early 2000's themed tech marketplace REST endpoints
  version: 1.0.0
  contact:
    name: API Support
    email: vertex4004@protonmail.com
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Express.js Dev & Prod
security:
  - bearerAuth: [ ]
tags:
  - name: ⬜ Health
    description: Health check
  - name: ⬜ Auth
    description: Authentication endpoints (register, login)
  - name: 🟥 Accounts
    description: Accounts, Users & Businesses
  - name: 🟦 Categories
    description: Categories, Attributes & ListingAttributeValues
  - name: 🟩 Listings
    description: Listings & ListingImages
  - name: 🟪 Orders
    description: Orders, OrderListings & CartItems
# endregion OAS3

# region COMPONENTS
components:
  # region Secutity Scheme
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  # endregion


  schemas:
    ErrorResponse:
      type: object
      description: Standardized error response format
      properties:
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Resource not found

    # region User Models/DTOs
    User (default):
      type: object
      description: Full user profile information including password
      properties:
        id:
          type: string
          example: uuid-here
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        createdAt:
          type: string
          format: date-time
          example: '1970-01-01T00:00:00Z'

    UserPublic:
      type: object
      description: Full user profile information (excluding password)
      properties:
        id:
          type: string
          example: uuid-here
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
    # endregion User Models/DTOs


  # region Error Responses
  responses:
    400BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericBadRequest:
              value:
                error: Bad request
                message: Invalid request format
            missingField:
              value:
                error: Bad request
                message: Missing required field
            invalidEmail:
              value:
                error: Bad request
                message: Invalid email format

    401Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericUnauthorized:
              value:
                error: Unauthorized
                message: Please provide valid authentication credentials

    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericNotFound:
              value:
                error: Not Found
                message: Resource not found
            userNotFound:
              value:
                error: Not Found
                message: User with id 123 not found

    409Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericConflict:
              value:
                error: Conflict
                message: Conflict occurred
            usernameTaken:
              value:
                error: Conflict
                message: Username is already taken
            emailTaken:
              value:
                error: Conflict
                message: Email is already taken

    500InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericServerError:
              value:
                error: Internal Server Error
                message: The server was unable to complete your request
  # endregion Error Responses
# endregion COMPONENTS



paths:
  # region PATHS
  # region ⬜ Health
  /health:
    get:
      tags:
        - ⬜ Health
      summary: Health check
      description: Returns a simple status object to indicate the API is reachable.
      operationId: getHealth
      responses:
        '200':
          description: 200 OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  timestamp:
                    type: string
                    format: date-time
                    example: '1970-01-01T00:00:00Z'
                  service:
                    type: string
                    example: Memory Mall e-commerce API
  # endregion ⬜ Health


  # region ⬜ Auth
  /auth/register/user:
    post:
      tags:
        - ⬜ Auth
      summary: Register new user
      description: Create a new user account and return a JWT token and basic user info.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  format: password
                  example: securePassword123
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                  user:
                    $ref: '#/components/schemas/UserPublic'
                  token:
                    type: string
                    example: jwt-token-here
        '400':
          $ref: '#/components/responses/400BadRequest'
        '409':
          $ref: '#/components/responses/409Conflict'
        '500':
          $ref: '#/components/responses/500InternalServerError'


  /auth/register/business:
    post:
      tags:
        - ⬜ Auth
      summary: Register new business
      description: Create a new business account and return a JWT token and basic user info.
      operationId: registerBusiness
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  format: password
                  example: securePassword123
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                  user:
                    $ref: '#/components/schemas/UserPublic'
                  token:
                    type: string
                    example: jwt-token-here
        '400':
          $ref: '#/components/responses/400BadRequest'
        '409':
          $ref: '#/components/responses/409Conflict'
        '500':
          $ref: '#/components/responses/500InternalServerError'


  /auth/login:
    post:
      tags:
        - ⬜ Auth
      summary: Login existing account (user/business)
      description: Authenticate an account (user/business) with email and password and return a JWT token.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  token:
                    type: string
                    example: jwt-token-here
                  user:
                    $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  # endregion ⬜ Auth
  # endregion PATHS